Code review, broke out into independent .js file
